NixOS configuration - testing, debugging, and organization
The aim of this PR is to contribute a guide to the topic of `NixOS configuration testing debugging and organization`.
It revises https://nix.dev/tutorials/integration-testing-using-virtual-machines and utilizes additional information from nix wiki and manuals.
It might happen that the topic of configuration file organization is split into a separate guide.
This initial commit has the function to show that I am working on the topic.

footnotes are mostly used to refer to the origin of parts of the documentation
many things are "todo" as addition or modification