Remove "reproducible" from glossary
It is not part of the official terminology the nixos documentation team established in https://github.com/NixOS/nix.dev/pull/263.

It was mixed in https://github.com/NixOS/nix.dev/pull/380 and messed up the official status of the terms. This change fixes that!

It can be added again when we figured out what we mean by it (https://github.com/NixOS/nix.dev/issues/464).