revert change of title
introducing a negatively connotated word is probably not a good idea.
there is some risk of inducing fear that was not there before, and the
tongue-in-cheek aspect may not come across.

the title should be unique and easy to recognise, and it was. "Reading
the Nix language" is already the title of the talk most of this is based
on, and we shouldn't reuse it