Update declarative-shell.md
Python is mentioned in the docs, but then the example uses Node JS. I am new to Nix, so I don't want to update the example to fully use Python, so I did the simpler change of fixing the heading and mention of Python to instead say Node JS.

If someone wants to fix the example code, or change the text to NodeJS or Node.JS, or whatever, by all means. My goal here is just to make the docs more clear. Thanks!