fix and simplify devmode script
previously, stderr was garbled, which made it really annoying to deal
with build errors.

it turns out that the library wrapper around calling a process is not
just counterproductive, but also completely unnecessary.

the build now takes longer because it's not incremental any more.
but the problem with `sphinx-build` in a shell environment is that it
doesn't keep track of an entire class of files, that would have to be
considered manually. this is too much manual overhead that doesn't scale
at all. with the new setup the live preview will do exactly what the
final deployment does.

[1]: https://github.com/executablebooks/sphinx-autobuild#relevant-sphinx-bugs

Co-authored-by: Alex Groleau source@proof.construction