update information flow diagram
we probably don't need it here in particular, but originally when the documentation team was new, it was useful to understand and communicate what's going on. it can probably stay here until we either find a better place or feel like it's in the way.

until then, a small update for clarity and consistency.