add a script to update Nixpkgs releases
this will accumulate releases over time. it's not great, but this is the
price of persistent URLs. otherwise, when people link to supposedly stable
URLs, and we garbage-collect old manuals, those links will rot away.

it's also slightly simpler algorithmically, and therefore easier to
reason about.

this can be incrementally improved in the future, by automatically
adding redirects to e.g. the closest following release for which the
manual is still served. for example, if we currently serve

    23.11
    23.05
    22.11
    22.05

but in the future decide to cut that list to the first two, the script
would drop the excess pins and add redirects like these:

    /manual/nixpkgs/22.05/* /manual/nixpkgs/23.05/:splat 301
    /manual/nixpkgs/22.11/* /manual/nixpkgs/23.05/:splat 301

the condition for that would be that we are reasonably sure that the
manuals will manage their own internal redirects as pages and anchors
move around between releases.