correct the number of flake users
The statistic in
https://discourse.nixos.org/t/nix-community-survey-2023-results/33124
does also include N/A responses in the percentage, so the 58% is a bit
misleading as the conclusion would be that 42% are still not using
flakes. I dropped the 1/3 statement for now, because I couldn't find any
information about that, but I can re-add it if this percentage is
actually correct. I also dropped the 2022 results as it didn't seem to
have the explicit question in it.