module system deep dive: minor improvements
Add more error checking to the `map.sh` and `geocode.sh` script

The default zoom of 2 results in the `center` option not showing anything
different. So, we change the default zoom to 10.

We also clear up the fact that using an option type of `nullOr <blah>` does not
set a default value of `null` for you.