Update towards-reproducibility-pinning-nixpkgs.md
I expected the `<nixpkgs>` link to point to a resource that explains what that construct is and not directly to Nixpkgs. If I would be totally new to Nix, that would be even more confusing.

Couldn't test whether the links in the new note section resolve properly as MyST wouldn't compile on my Mac M1. If you you like this change, I can spend more time to test them properly.