make building the manuals optional
since we currently can't fetch the manuals as a tarball from Hydra, we
have to evaluate all the Nix releases under the sun, which takes quite
a while. most of the time this is not needed and is in the way to work
on the site, so disable it by default.