Publish PDF
I had to update pinned nixpkgs, as it did not build with TexLive 2022.

Some content annotations are not transformed very well in LaTex output. For example, there is a [bug](https://github.com/executablebooks/sphinx-design/issues/107) in sphinx-design which prints dropdowns just as subsections without any visual highlighting. In general it is readable and better typeset compared to the PDF obtained by printing to PDF.

Also, there is now full texlive which is quite a huge dependency.