nixos-configuration-on-vm: de-emphasize `nixos-generate-config`
The main point of `nixos-generate-config` is that it generates the
`hardware-configuration.nix` matching the characteristics of the machine
you're running on, but that configuration is not used in the VM scenario
anyway.

I think it distracts from the main message of this tutorial, especially
as there's some outstanding bugs in `nixos-generate-config` that you
might get confused by, such as https://github.com/NixOS/nixpkgs/issues/417803

I'd suggest removing the `nixos-generate-config` content from this
tutorial. Possibly we could have it return as a dedicated tutorial at
some point.